Removed imports examples and renamed leoNodes to leoOutlineNodes
