--- src/leo/app/leoApp.js
@@ -1,5 +1,0 @@
-function test2(info) {
-    info("leoApp:test2");
-}
-// export {test2};
-module.exports = {test2};
