+++ src/leojs.ts
@@ -1,11 +1,4 @@
 
-import * as vscode from 'vscode';
-
-// import * as abc from './leoFiles'; // works
-import { test, testVar } from './leoFiles'; // works too
-
-import * as leo from './leo/leo'; // Your tests from leojs4 :)
-
 import { LeoButton, LeoDocument, PNode } from './types';
 
 // https://github.com/leo-editor/leo-editor/issues/1025
@@ -78,16 +71,6 @@ export class Leojs {
             // set to undefined at first for root nodes
             this._setPNodeParents(p_position, undefined);
         });
-
-
-
-        // abc.test("the string is " + abc.testVar);
-        test("the string is " + testVar);
-
-        const info = vscode.window.showInformationMessage;
-        leo.test(info);
-
-
     }
 
     /**
